requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.1
pandas==1.3.1
Pillow==8.3.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.1
requests==2.26.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
